+++ tests/test_manufacturer_view.py
